Crate google_container1

This documentation was generated from Container crate version 1.0.9+20190610, where 20190610 is the exact revision of the container:v1 schema built by the mako code generator v1.0.9.

Everything else about the Container v1 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

projects

aggregated usable subnetworks list, locations clusters complete ip rotation, locations clusters create, locations clusters delete, locations clusters get, locations clusters get jwks, locations clusters list, locations clusters node pools create, locations clusters node pools delete, locations clusters node pools get, locations clusters node pools list, locations clusters node pools rollback, locations clusters node pools set autoscaling, locations clusters node pools set management, locations clusters node pools set size, locations clusters node pools update, locations clusters set addons, locations clusters set legacy abac, locations clusters set locations, locations clusters set logging, locations clusters set maintenance policy, locations clusters set master auth, locations clusters set monitoring, locations clusters set network policy, locations clusters set resource labels, locations clusters start ip rotation, locations clusters update, locations clusters update master, locations clusters well-known get openid-configuration, locations get server config, locations operations cancel, locations operations get, locations operations list, zones clusters addons, zones clusters complete ip rotation, zones clusters create, zones clusters delete, zones clusters get, zones clusters legacy abac, zones clusters list, zones clusters locations, zones clusters logging, zones clusters master, zones clusters monitoring, zones clusters node pools autoscaling, zones clusters node pools create, zones clusters node pools delete, zones clusters node pools get, zones clusters node pools list, zones clusters node pools rollback, zones clusters node pools set management, zones clusters node pools set size, zones clusters node pools update, zones clusters resource labels, zones clusters set maintenance policy, zones clusters set master auth, zones clusters set network policy, zones clusters start ip rotation, zones clusters update, zones get serverconfig, zones operations cancel, zones operations get and zones operations list

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.projects().zones_clusters_node_pools_rollback(...).doit()
let r = hub.projects().locations_clusters_node_pools_set_autoscaling(...).doit()
let r = hub.projects().zones_clusters_node_pools_create(...).doit()
let r = hub.projects().zones_clusters_monitoring(...).doit()
let r = hub.projects().locations_clusters_update(...).doit()
let r = hub.projects().zones_clusters_delete(...).doit()
let r = hub.projects().zones_clusters_legacy_abac(...).doit()
let r = hub.projects().zones_clusters_addons(...).doit()
let r = hub.projects().locations_clusters_delete(...).doit()
let r = hub.projects().locations_clusters_node_pools_delete(...).doit()
let r = hub.projects().zones_clusters_start_ip_rotation(...).doit()
let r = hub.projects().zones_clusters_create(...).doit()
let r = hub.projects().locations_clusters_set_addons(...).doit()
let r = hub.projects().locations_clusters_complete_ip_rotation(...).doit()
let r = hub.projects().locations_clusters_node_pools_update(...).doit()
let r = hub.projects().locations_clusters_node_pools_set_size(...).doit()
let r = hub.projects().locations_clusters_set_network_policy(...).doit()
let r = hub.projects().zones_clusters_node_pools_set_management(...).doit()
let r = hub.projects().zones_operations_get(...).doit()
let r = hub.projects().locations_clusters_set_monitoring(...).doit()
let r = hub.projects().locations_clusters_set_locations(...).doit()
let r = hub.projects().zones_clusters_logging(...).doit()
let r = hub.projects().zones_clusters_update(...).doit()
let r = hub.projects().zones_clusters_node_pools_autoscaling(...).doit()
let r = hub.projects().zones_clusters_node_pools_set_size(...).doit()
let r = hub.projects().locations_clusters_set_legacy_abac(...).doit()
let r = hub.projects().locations_clusters_create(...).doit()
let r = hub.projects().locations_clusters_node_pools_set_management(...).doit()
let r = hub.projects().locations_clusters_update_master(...).doit()
let r = hub.projects().zones_clusters_set_maintenance_policy(...).doit()
let r = hub.projects().locations_clusters_node_pools_rollback(...).doit()
let r = hub.projects().zones_clusters_node_pools_delete(...).doit()
let r = hub.projects().zones_clusters_locations(...).doit()
let r = hub.projects().locations_clusters_start_ip_rotation(...).doit()
let r = hub.projects().locations_clusters_set_resource_labels(...).doit()
let r = hub.projects().locations_clusters_node_pools_create(...).doit()
let r = hub.projects().zones_clusters_master(...).doit()
let r = hub.projects().zones_clusters_node_pools_update(...).doit()
let r = hub.projects().zones_clusters_set_network_policy(...).doit()
let r = hub.projects().locations_clusters_set_maintenance_policy(...).doit()
let r = hub.projects().zones_clusters_set_master_auth(...).doit()
let r = hub.projects().locations_clusters_set_master_auth(...).doit()
let r = hub.projects().zones_clusters_complete_ip_rotation(...).doit()
let r = hub.projects().zones_clusters_resource_labels(...).doit()
let r = hub.projects().locations_clusters_set_logging(...).doit()
let r = hub.projects().locations_operations_get(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-container1 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.10"
hyper-rustls = "^0.6"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^1.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_container1 as container1;
use container1::RollbackNodePoolUpgradeRequest;
use container1::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use container1::Container;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Container::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = RollbackNodePoolUpgradeRequest::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.projects().zones_clusters_node_pools_rollback(req, "projectId", "zone", "clusterId", "nodePoolId")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

AcceleratorConfig
AcceleratorConfig represents a Hardware Accelerator request.

AddonsConfig
Configuration for the addons that can be automatically spun up in the cluster, enabling additional functionality.

AutoUpgradeOptions
AutoUpgradeOptions defines the set of options for the user to control how the Auto Upgrades will proceed.

BigQueryDestination
Parameters for using BigQuery as the destination of resource usage export.

CancelOperationRequest
CancelOperationRequest cancels a single operation.

Chunk

CidrBlock
CidrBlock contains an optional name and one CIDR block.

ClientCertificateConfig
Configuration for client certificates on the cluster.

Cluster
A Google Kubernetes Engine cluster.

ClusterUpdate
ClusterUpdate describes an update to the cluster. Exactly one update can be applied to a cluster with each request, so at most one field can be provided.

CompleteIPRotationRequest
CompleteIPRotationRequest moves the cluster master back into single-IP mode.

ConsumptionMeteringConfig
Parameters for controlling consumption metering.

Container
Central instance to access all Container related resource activities

ContentRange
Implements the Content-Range header, for serialization only

CreateClusterRequest
CreateClusterRequest creates a cluster.

CreateNodePoolRequest
CreateNodePoolRequest creates a node pool for a cluster.

DailyMaintenanceWindow
Time window specified for daily maintenance operations.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

DummyNetworkStream

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance:

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

GetJSONWebKeysResponse
GetJSONWebKeysResponse is a valid JSON Web Key Set as specififed in rfc 7517

GetOpenIDConfigResponse
GetOpenIDConfigResponse is an OIDC discovery document for the cluster. See the OpenID Connect Discovery 1.0 specification for details.

HorizontalPodAutoscaling
Configuration options for the horizontal pod autoscaling feature, which increases or decreases the number of replica pods a replication controller has based on the resource usage of the existing pods.

HttpLoadBalancing
Configuration options for the HTTP (L7) load balancing controller addon, which makes it easy to set up HTTP load balancers for services in a cluster.

IPAllocationPolicy
Configuration for controlling how IPs are allocated in the cluster.

JsonServerError
A utility type which can decode a server response that indicates error

Jwk
Jwk is a JSON Web Key as specified in RFC 7517

KubernetesDashboard
Configuration for the Kubernetes Dashboard.

LegacyAbac
Configuration for the legacy Attribute Based Access Control authorization mode.

ListClustersResponse
ListClustersResponse is the result of ListClustersRequest.

ListNodePoolsResponse
ListNodePoolsResponse is the result of ListNodePoolsRequest.

ListOperationsResponse
ListOperationsResponse is the result of ListOperationsRequest.

ListUsableSubnetworksResponse
ListUsableSubnetworksResponse is the response of ListUsableSubnetworksRequest.

MaintenancePolicy
MaintenancePolicy defines the maintenance policy to be used for the cluster.

MaintenanceWindow
MaintenanceWindow defines the maintenance window to be used for the cluster.

MasterAuth
The authentication information for accessing the master endpoint. Authentication can be done using HTTP basic auth or using client certificates.

MasterAuthorizedNetworksConfig
Configuration options for the master authorized networks feature. Enabled master authorized networks will disallow all external traffic to access Kubernetes master through HTTPS except traffic from the given CIDR blocks, Google Compute Engine Public IPs and Google Prod IPs.

MaxPodsConstraint
Constraints applied to pods.

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

NetworkConfig
NetworkConfig reports the relative names of network & subnetwork.

NetworkPolicy
Configuration options for the NetworkPolicy feature. https://kubernetes.io/docs/concepts/services-networking/networkpolicies/

NetworkPolicyConfig
Configuration for NetworkPolicy. This only tracks whether the addon is enabled or not on the Master, it does not track whether network policy is enabled for the nodes.

NodeConfig
Parameters that describe the nodes in a cluster.

NodeManagement
NodeManagement defines the set of node management services turned on for the node pool.

NodePool
NodePool contains the name and configuration for a cluster's node pool. Node pools are a set of nodes (i.e. VM's), with a common configuration and specification, under the control of the cluster master. They may have a set of Kubernetes labels applied to them, which may be used to reference them during pod scheduling. They may also be resized up or down, to accommodate the workload.

NodePoolAutoscaling
NodePoolAutoscaling contains information required by cluster autoscaler to adjust the size of the node pool to the current cluster usage.

NodeTaint
Kubernetes taint is comprised of three fields: key, value, and effect. Effect can only be one of three types: NoSchedule, PreferNoSchedule or NoExecute.

Operation
This operation resource represents operations that may have happened or are happening on the cluster. All fields are output only.

PrivateClusterConfig
Configuration options for private clusters.

ProjectAggregatedUsableSubnetworkListCall
Lists subnetworks that are usable for creating clusters in a project.

ProjectLocationClusterCompleteIpRotationCall
Completes master IP rotation.

ProjectLocationClusterCreateCall
Creates a cluster, consisting of the specified number and type of Google Compute Engine instances.

ProjectLocationClusterDeleteCall
Deletes the cluster, including the Kubernetes endpoint and all worker nodes.

ProjectLocationClusterGetCall
Gets the details of a specific cluster.

ProjectLocationClusterGetJwkCall
Gets the public component of the cluster signing keys in JSON Web Key format. This API is not yet intended for general use, and is not available for all clusters.

ProjectLocationClusterListCall
Lists all clusters owned by a project in either the specified zone or all zones.

ProjectLocationClusterNodePoolCreateCall
Creates a node pool for a cluster.

ProjectLocationClusterNodePoolDeleteCall
Deletes a node pool from a cluster.

ProjectLocationClusterNodePoolGetCall
Retrieves the requested node pool.

ProjectLocationClusterNodePoolListCall
Lists the node pools for a cluster.

ProjectLocationClusterNodePoolRollbackCall
Rolls back a previously Aborted or Failed NodePool upgrade. This makes no changes if the last upgrade successfully completed.

ProjectLocationClusterNodePoolSetAutoscalingCall
Sets the autoscaling settings for the specified node pool.

ProjectLocationClusterNodePoolSetManagementCall
Sets the NodeManagement options for a node pool.

ProjectLocationClusterNodePoolSetSizeCall
Sets the size for a specific node pool.

ProjectLocationClusterNodePoolUpdateCall
Updates the version and/or image type for the specified node pool.

ProjectLocationClusterSetAddonCall
Sets the addons for a specific cluster.

ProjectLocationClusterSetLegacyAbacCall
Enables or disables the ABAC authorization mechanism on a cluster.

ProjectLocationClusterSetLocationCall
Sets the locations for a specific cluster.

ProjectLocationClusterSetLoggingCall
Sets the logging service for a specific cluster.

ProjectLocationClusterSetMaintenancePolicyCall
Sets the maintenance policy for a cluster.

ProjectLocationClusterSetMasterAuthCall
Sets master auth materials. Currently supports changing the admin password or a specific cluster, either via password generation or explicitly setting the password.

ProjectLocationClusterSetMonitoringCall
Sets the monitoring service for a specific cluster.

ProjectLocationClusterSetNetworkPolicyCall
Enables or disables Network Policy for a cluster.

ProjectLocationClusterSetResourceLabelCall
Sets labels on a cluster.

ProjectLocationClusterStartIpRotationCall
Starts master IP rotation.

ProjectLocationClusterUpdateCall
Updates the settings of a specific cluster.

ProjectLocationClusterUpdateMasterCall
Updates the master for a specific cluster.

ProjectLocationClusterWellKnownGetOpenidConfigurationCall
Gets the OIDC discovery document for the cluster. See the OpenID Connect Discovery 1.0 specification for details. This API is not yet intended for general use, and is not available for all clusters.

ProjectLocationGetServerConfigCall
Returns configuration info about the Google Kubernetes Engine service.

ProjectLocationOperationCancelCall
Cancels the specified operation.

ProjectLocationOperationGetCall
Gets the specified operation.

ProjectLocationOperationListCall
Lists all operations in a project in a specific zone or all zones.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the Container hub.

ProjectZoneClusterAddonCall
Sets the addons for a specific cluster.

ProjectZoneClusterCompleteIpRotationCall
Completes master IP rotation.

ProjectZoneClusterCreateCall
Creates a cluster, consisting of the specified number and type of Google Compute Engine instances.

ProjectZoneClusterDeleteCall
Deletes the cluster, including the Kubernetes endpoint and all worker nodes.

ProjectZoneClusterGetCall
Gets the details of a specific cluster.

ProjectZoneClusterLegacyAbacCall
Enables or disables the ABAC authorization mechanism on a cluster.

ProjectZoneClusterListCall
Lists all clusters owned by a project in either the specified zone or all zones.

ProjectZoneClusterLocationCall
Sets the locations for a specific cluster.

ProjectZoneClusterLoggingCall
Sets the logging service for a specific cluster.

ProjectZoneClusterMasterCall
Updates the master for a specific cluster.

ProjectZoneClusterMonitoringCall
Sets the monitoring service for a specific cluster.

ProjectZoneClusterNodePoolAutoscalingCall
Sets the autoscaling settings for the specified node pool.

ProjectZoneClusterNodePoolCreateCall
Creates a node pool for a cluster.

ProjectZoneClusterNodePoolDeleteCall
Deletes a node pool from a cluster.

ProjectZoneClusterNodePoolGetCall
Retrieves the requested node pool.

ProjectZoneClusterNodePoolListCall
Lists the node pools for a cluster.

ProjectZoneClusterNodePoolRollbackCall
Rolls back a previously Aborted or Failed NodePool upgrade. This makes no changes if the last upgrade successfully completed.

ProjectZoneClusterNodePoolSetManagementCall
Sets the NodeManagement options for a node pool.

ProjectZoneClusterNodePoolSetSizeCall
Sets the size for a specific node pool.

ProjectZoneClusterNodePoolUpdateCall
Updates the version and/or image type for the specified node pool.

ProjectZoneClusterResourceLabelCall
Sets labels on a cluster.

ProjectZoneClusterSetMaintenancePolicyCall
Sets the maintenance policy for a cluster.

ProjectZoneClusterSetMasterAuthCall
Sets master auth materials. Currently supports changing the admin password or a specific cluster, either via password generation or explicitly setting the password.

ProjectZoneClusterSetNetworkPolicyCall
Enables or disables Network Policy for a cluster.

ProjectZoneClusterStartIpRotationCall
Starts master IP rotation.

ProjectZoneClusterUpdateCall
Updates the settings of a specific cluster.

ProjectZoneGetServerconfigCall
Returns configuration info about the Google Kubernetes Engine service.

ProjectZoneOperationCancelCall
Cancels the specified operation.

ProjectZoneOperationGetCall
Gets the specified operation.

ProjectZoneOperationListCall
Lists all operations in a project in a specific zone or all zones.

RangeResponseHeader

ResourceUsageExportConfig
Configuration for exporting cluster resource usages.

ResumableUploadHelper
A utility type to perform a resumable upload from start to end.

RollbackNodePoolUpgradeRequest
RollbackNodePoolUpgradeRequest rollbacks the previously Aborted or Failed NodePool upgrade. This will be an no-op if the last upgrade successfully completed.

ServerConfig
Kubernetes Engine service configuration.

ServerError

ServerMessage

SetAddonsConfigRequest
SetAddonsConfigRequest sets the addons associated with the cluster.

SetLabelsRequest
SetLabelsRequest sets the Google Cloud Platform labels on a Google Container Engine cluster, which will in turn set them for Google Compute Engine resources used by that cluster

SetLegacyAbacRequest
SetLegacyAbacRequest enables or disables the ABAC authorization mechanism for a cluster.

SetLocationsRequest
SetLocationsRequest sets the locations of the cluster.

SetLoggingServiceRequest
SetLoggingServiceRequest sets the logging service of a cluster.

SetMaintenancePolicyRequest
SetMaintenancePolicyRequest sets the maintenance policy for a cluster.

SetMasterAuthRequest
SetMasterAuthRequest updates the admin password of a cluster.

SetMonitoringServiceRequest
SetMonitoringServiceRequest sets the monitoring service of a cluster.

SetNetworkPolicyRequest
SetNetworkPolicyRequest enables/disables network policy for a cluster.

SetNodePoolAutoscalingRequest
SetNodePoolAutoscalingRequest sets the autoscaler settings of a node pool.

SetNodePoolManagementRequest
SetNodePoolManagementRequest sets the node management properties of a node pool.

SetNodePoolSizeRequest
SetNodePoolSizeRequest sets the size a node pool.

StartIPRotationRequest
StartIPRotationRequest creates a new IP for the cluster and then performs a node upgrade on each node pool to point to the new IP.

StatusCondition
StatusCondition describes why a cluster or a node pool has a certain status (e.g., ERROR or DEGRADED).

UpdateClusterRequest
UpdateClusterRequest updates the settings of a cluster.

UpdateMasterRequest
UpdateMasterRequest updates the master of the cluster.

UpdateNodePoolRequest
UpdateNodePoolRequests update a node pool's image and/or version.

UsableSubnetwork
UsableSubnetwork resource returns the subnetwork name, its associated network and the primary CIDR range.

UsableSubnetworkSecondaryRange
Secondary IP range of a usable subnetwork.

XUploadContentType
The X-Upload-Content-Type header.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.