Module google_container1::api

Structs

AcceleratorConfig
AcceleratorConfig represents a Hardware Accelerator request.

AddonsConfig
Configuration for the addons that can be automatically spun up in the cluster, enabling additional functionality.

AuthenticatorGroupsConfig
Configuration for returning group information from authenticators.

AutoUpgradeOptions
AutoUpgradeOptions defines the set of options for the user to control how the Auto Upgrades will proceed.

Autopilot
Autopilot is the configuration for Autopilot settings on the cluster.

AutoprovisioningNodePoolDefaults
AutoprovisioningNodePoolDefaults contains defaults for a node pool created by NAP.

BigQueryDestination
Parameters for using BigQuery as the destination of resource usage export.

BinaryAuthorization
Configuration for Binary Authorization.

CancelOperationRequest
CancelOperationRequest cancels a single operation.

CidrBlock
CidrBlock contains an optional name and one CIDR block.

ClientCertificateConfig
Configuration for client certificates on the cluster.

CloudRunConfig
Configuration options for the Cloud Run feature.

Cluster
A Google Kubernetes Engine cluster.

ClusterAutoscaling
ClusterAutoscaling contains global, per-cluster information required by Cluster Autoscaler to automatically adjust the size of the cluster and create/delete node pools based on the current needs.

ClusterUpdate
ClusterUpdate describes an update to the cluster. Exactly one update can be applied to a cluster with each request, so at most one field can be provided.

CompleteIPRotationRequest
CompleteIPRotationRequest moves the cluster master back into single-IP mode.

ConfigConnectorConfig
Configuration options for the Config Connector add-on.

ConsumptionMeteringConfig
Parameters for controlling consumption metering.

Container
Central instance to access all Container related resource activities

CreateClusterRequest
CreateClusterRequest creates a cluster.

CreateNodePoolRequest
CreateNodePoolRequest creates a node pool for a cluster.

DailyMaintenanceWindow
Time window specified for daily maintenance operations.

DatabaseEncryption
Configuration of etcd encryption.

DefaultSnatStatus
DefaultSnatStatus contains the desired state of whether default sNAT should be disabled on the cluster.

DnsCacheConfig
Configuration for NodeLocal DNSCache

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

GcePersistentDiskCsiDriverConfig
Configuration for the Compute Engine PD CSI driver.

GetJSONWebKeysResponse
GetJSONWebKeysResponse is a valid JSON Web Key Set as specififed in rfc 7517

GetOpenIDConfigResponse
GetOpenIDConfigResponse is an OIDC discovery document for the cluster. See the OpenID Connect Discovery 1.0 specification for details.

HorizontalPodAutoscaling
Configuration options for the horizontal pod autoscaling feature, which increases or decreases the number of replica pods a replication controller has based on the resource usage of the existing pods.

HttpCacheControlResponseHeader
RFC-2616: cache control support

HttpLoadBalancing
Configuration options for the HTTP (L7) load balancing controller addon, which makes it easy to set up HTTP load balancers for services in a cluster.

IPAllocationPolicy
Configuration for controlling how IPs are allocated in the cluster.

IntraNodeVisibilityConfig
IntraNodeVisibilityConfig contains the desired config of the intra-node visibility on this cluster.

Jwk
Jwk is a JSON Web Key as specified in RFC 7517

KubernetesDashboard
Configuration for the Kubernetes Dashboard.

LegacyAbac
Configuration for the legacy Attribute Based Access Control authorization mode.

LinuxNodeConfig
Parameters that can be configured on Linux nodes.

ListClustersResponse
ListClustersResponse is the result of ListClustersRequest.

ListNodePoolsResponse
ListNodePoolsResponse is the result of ListNodePoolsRequest.

ListOperationsResponse
ListOperationsResponse is the result of ListOperationsRequest.

ListUsableSubnetworksResponse
ListUsableSubnetworksResponse is the response of ListUsableSubnetworksRequest.

MaintenancePolicy
MaintenancePolicy defines the maintenance policy to be used for the cluster.

MaintenanceWindow
MaintenanceWindow defines the maintenance window to be used for the cluster.

MasterAuth
The authentication information for accessing the master endpoint. Authentication can be done using HTTP basic auth or using client certificates.

MasterAuthorizedNetworksConfig
Configuration options for the master authorized networks feature. Enabled master authorized networks will disallow all external traffic to access Kubernetes master through HTTPS except traffic from the given CIDR blocks, Google Compute Engine Public IPs and Google Prod IPs.

MaxPodsConstraint
Constraints applied to pods.

Metric
Progress metric is (string, int|float|string) pair.

NetworkConfig
NetworkConfig reports the relative names of network & subnetwork.

NetworkPolicy
Configuration options for the NetworkPolicy feature. https://kubernetes.io/docs/concepts/services-networking/networkpolicies/

NetworkPolicyConfig
Configuration for NetworkPolicy. This only tracks whether the addon is enabled or not on the Master, it does not track whether network policy is enabled for the nodes.

NodeConfig
Parameters that describe the nodes in a cluster.

NodeKubeletConfig
Node kubelet configs.

NodeManagement
NodeManagement defines the set of node management services turned on for the node pool.

NodePool
NodePool contains the name and configuration for a cluster's node pool. Node pools are a set of nodes (i.e. VM's), with a common configuration and specification, under the control of the cluster master. They may have a set of Kubernetes labels applied to them, which may be used to reference them during pod scheduling. They may also be resized up or down, to accommodate the workload.

NodePoolAutoscaling
NodePoolAutoscaling contains information required by cluster autoscaler to adjust the size of the node pool to the current cluster usage.

NodeTaint
Kubernetes taint is comprised of three fields: key, value, and effect. Effect can only be one of three types: NoSchedule, PreferNoSchedule or NoExecute. See here for more information, including usage and the valid values.

NotificationConfig
NotificationConfig is the configuration of notifications.

Operation
This operation resource represents operations that may have happened or are happening on the cluster. All fields are output only.

OperationProgress
Information about operation (or operation stage) progress.

PrivateClusterConfig
Configuration options for private clusters.

PrivateClusterMasterGlobalAccessConfig
Configuration for controlling master global access settings.

ProjectAggregatedUsableSubnetworkListCall
Lists subnetworks that are usable for creating clusters in a project.

ProjectLocationClusterCompleteIpRotationCall
Completes master IP rotation.

ProjectLocationClusterCreateCall
Creates a cluster, consisting of the specified number and type of Google Compute Engine instances. By default, the cluster is created in the project's default network. One firewall is added for the cluster. After cluster creation, the Kubelet creates routes for each node to allow the containers on that node to communicate with all other instances in the cluster. Finally, an entry is added to the project's global metadata indicating which CIDR range the cluster is using.

ProjectLocationClusterDeleteCall
Deletes the cluster, including the Kubernetes endpoint and all worker nodes. Firewalls and routes that were configured during cluster creation are also deleted. Other Google Compute Engine resources that might be in use by the cluster, such as load balancer resources, are not deleted if they weren't present when the cluster was initially created.

ProjectLocationClusterGetCall
Gets the details of a specific cluster.

ProjectLocationClusterGetJwkCall
Gets the public component of the cluster signing keys in JSON Web Key format. This API is not yet intended for general use, and is not available for all clusters.

ProjectLocationClusterListCall
Lists all clusters owned by a project in either the specified zone or all zones.

ProjectLocationClusterNodePoolCreateCall
Creates a node pool for a cluster.

ProjectLocationClusterNodePoolDeleteCall
Deletes a node pool from a cluster.

ProjectLocationClusterNodePoolGetCall
Retrieves the requested node pool.

ProjectLocationClusterNodePoolListCall
Lists the node pools for a cluster.

ProjectLocationClusterNodePoolRollbackCall
Rolls back a previously Aborted or Failed NodePool upgrade. This makes no changes if the last upgrade successfully completed.

ProjectLocationClusterNodePoolSetAutoscalingCall
Sets the autoscaling settings for the specified node pool.

ProjectLocationClusterNodePoolSetManagementCall
Sets the NodeManagement options for a node pool.

ProjectLocationClusterNodePoolSetSizeCall
Sets the size for a specific node pool. The new size will be used for all replicas, including future replicas created by modifying NodePool.locations.

ProjectLocationClusterNodePoolUpdateCall
Updates the version and/or image type for the specified node pool.

ProjectLocationClusterSetAddonCall
Sets the addons for a specific cluster.

ProjectLocationClusterSetLegacyAbacCall
Enables or disables the ABAC authorization mechanism on a cluster.

ProjectLocationClusterSetLocationCall
Sets the locations for a specific cluster. Deprecated. Use projects.locations.clusters.update instead.

ProjectLocationClusterSetLoggingCall
Sets the logging service for a specific cluster.

ProjectLocationClusterSetMaintenancePolicyCall
Sets the maintenance policy for a cluster.

ProjectLocationClusterSetMasterAuthCall
Sets master auth materials. Currently supports changing the admin password or a specific cluster, either via password generation or explicitly setting the password.

ProjectLocationClusterSetMonitoringCall
Sets the monitoring service for a specific cluster.

ProjectLocationClusterSetNetworkPolicyCall
Enables or disables Network Policy for a cluster.

ProjectLocationClusterSetResourceLabelCall
Sets labels on a cluster.

ProjectLocationClusterStartIpRotationCall
Starts master IP rotation.

ProjectLocationClusterUpdateCall
Updates the settings of a specific cluster.

ProjectLocationClusterUpdateMasterCall
Updates the master for a specific cluster.

ProjectLocationClusterWellKnownGetOpenidConfigurationCall
Gets the OIDC discovery document for the cluster. See the OpenID Connect Discovery 1.0 specification for details. This API is not yet intended for general use, and is not available for all clusters.

ProjectLocationGetServerConfigCall
Returns configuration info about the Google Kubernetes Engine service.

ProjectLocationOperationCancelCall
Cancels the specified operation.

ProjectLocationOperationGetCall
Gets the specified operation.

ProjectLocationOperationListCall
Lists all operations in a project in a specific zone or all zones.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the Container hub.

ProjectZoneClusterAddonCall
Sets the addons for a specific cluster.

ProjectZoneClusterCompleteIpRotationCall
Completes master IP rotation.

ProjectZoneClusterCreateCall
Creates a cluster, consisting of the specified number and type of Google Compute Engine instances. By default, the cluster is created in the project's default network. One firewall is added for the cluster. After cluster creation, the Kubelet creates routes for each node to allow the containers on that node to communicate with all other instances in the cluster. Finally, an entry is added to the project's global metadata indicating which CIDR range the cluster is using.

ProjectZoneClusterDeleteCall
Deletes the cluster, including the Kubernetes endpoint and all worker nodes. Firewalls and routes that were configured during cluster creation are also deleted. Other Google Compute Engine resources that might be in use by the cluster, such as load balancer resources, are not deleted if they weren't present when the cluster was initially created.

ProjectZoneClusterGetCall
Gets the details of a specific cluster.

ProjectZoneClusterLegacyAbacCall
Enables or disables the ABAC authorization mechanism on a cluster.

ProjectZoneClusterListCall
Lists all clusters owned by a project in either the specified zone or all zones.

ProjectZoneClusterLocationCall
Sets the locations for a specific cluster. Deprecated. Use projects.locations.clusters.update instead.

ProjectZoneClusterLoggingCall
Sets the logging service for a specific cluster.

ProjectZoneClusterMasterCall
Updates the master for a specific cluster.

ProjectZoneClusterMonitoringCall
Sets the monitoring service for a specific cluster.

ProjectZoneClusterNodePoolAutoscalingCall
Sets the autoscaling settings for the specified node pool.

ProjectZoneClusterNodePoolCreateCall
Creates a node pool for a cluster.

ProjectZoneClusterNodePoolDeleteCall
Deletes a node pool from a cluster.

ProjectZoneClusterNodePoolGetCall
Retrieves the requested node pool.

ProjectZoneClusterNodePoolListCall
Lists the node pools for a cluster.

ProjectZoneClusterNodePoolRollbackCall
Rolls back a previously Aborted or Failed NodePool upgrade. This makes no changes if the last upgrade successfully completed.

ProjectZoneClusterNodePoolSetManagementCall
Sets the NodeManagement options for a node pool.

ProjectZoneClusterNodePoolSetSizeCall
Sets the size for a specific node pool. The new size will be used for all replicas, including future replicas created by modifying NodePool.locations.

ProjectZoneClusterNodePoolUpdateCall
Updates the version and/or image type for the specified node pool.

ProjectZoneClusterResourceLabelCall
Sets labels on a cluster.

ProjectZoneClusterSetMaintenancePolicyCall
Sets the maintenance policy for a cluster.

ProjectZoneClusterSetMasterAuthCall
Sets master auth materials. Currently supports changing the admin password or a specific cluster, either via password generation or explicitly setting the password.

ProjectZoneClusterSetNetworkPolicyCall
Enables or disables Network Policy for a cluster.

ProjectZoneClusterStartIpRotationCall
Starts master IP rotation.

ProjectZoneClusterUpdateCall
Updates the settings of a specific cluster.

ProjectZoneGetServerconfigCall
Returns configuration info about the Google Kubernetes Engine service.

ProjectZoneOperationCancelCall
Cancels the specified operation.

ProjectZoneOperationGetCall
Gets the specified operation.

ProjectZoneOperationListCall
Lists all operations in a project in a specific zone or all zones.

PubSub
Pub/Sub specific notification config.

RecurringTimeWindow
Represents an arbitrary window of time that recurs.

ReleaseChannel
ReleaseChannel indicates which release channel a cluster is subscribed to. Release channels are arranged in order of risk. When a cluster is subscribed to a release channel, Google maintains both the master version and the node version. Node auto-upgrade defaults to true and cannot be disabled.

ReleaseChannelConfig
ReleaseChannelConfig exposes configuration for a release channel.

ReservationAffinity
ReservationAffinity is the configuration of desired reservation which instances could take capacity from.

ResourceLimit
Contains information about amount of some resource in the cluster. For memory, value should be in GB.

ResourceUsageExportConfig
Configuration for exporting cluster resource usages.

RollbackNodePoolUpgradeRequest
RollbackNodePoolUpgradeRequest rollbacks the previously Aborted or Failed NodePool upgrade. This will be an no-op if the last upgrade successfully completed.

SandboxConfig
SandboxConfig contains configurations of the sandbox to use for the node.

ServerConfig
Kubernetes Engine service configuration.

SetAddonsConfigRequest
SetAddonsConfigRequest sets the addons associated with the cluster.

SetLabelsRequest
SetLabelsRequest sets the Google Cloud Platform labels on a Google Container Engine cluster, which will in turn set them for Google Compute Engine resources used by that cluster

SetLegacyAbacRequest
SetLegacyAbacRequest enables or disables the ABAC authorization mechanism for a cluster.

SetLocationsRequest
SetLocationsRequest sets the locations of the cluster.

SetLoggingServiceRequest
SetLoggingServiceRequest sets the logging service of a cluster.

SetMaintenancePolicyRequest
SetMaintenancePolicyRequest sets the maintenance policy for a cluster.

SetMasterAuthRequest
SetMasterAuthRequest updates the admin password of a cluster.

SetMonitoringServiceRequest
SetMonitoringServiceRequest sets the monitoring service of a cluster.

SetNetworkPolicyRequest
SetNetworkPolicyRequest enables/disables network policy for a cluster.

SetNodePoolAutoscalingRequest
SetNodePoolAutoscalingRequest sets the autoscaler settings of a node pool.

SetNodePoolManagementRequest
SetNodePoolManagementRequest sets the node management properties of a node pool.

SetNodePoolSizeRequest
SetNodePoolSizeRequest sets the size of a node pool.

ShieldedInstanceConfig
A set of Shielded Instance options.

ShieldedNodes
Configuration of Shielded Nodes feature.

StartIPRotationRequest
StartIPRotationRequest creates a new IP for the cluster and then performs a node upgrade on each node pool to point to the new IP.

Status
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the API Design Guide.

StatusCondition
StatusCondition describes why a cluster or a node pool has a certain status (e.g., ERROR or DEGRADED).

TimeWindow
Represents an arbitrary window of time.

UpdateClusterRequest
UpdateClusterRequest updates the settings of a cluster.

UpdateMasterRequest
UpdateMasterRequest updates the master of the cluster.

UpdateNodePoolRequest
UpdateNodePoolRequests update a node pool's image and/or version.

UpgradeSettings
These upgrade settings control the level of parallelism and the level of disruption caused by an upgrade. maxUnavailable controls the number of nodes that can be simultaneously unavailable. maxSurge controls the number of additional nodes that can be added to the node pool temporarily for the time of the upgrade to increase the number of available nodes. (maxUnavailable + maxSurge) determines the level of parallelism (how many nodes are being upgraded at the same time). Note: upgrades inevitably introduce some disruption since workloads need to be moved from old nodes to new, upgraded ones. Even if maxUnavailable=0, this holds true. (Disruption stays within the limits of PodDisruptionBudget, if it is configured.) Consider a hypothetical node pool with 5 nodes having maxSurge=2, maxUnavailable=1. This means the upgrade process upgrades 3 nodes simultaneously. It creates 2 additional (upgraded) nodes, then it brings down 3 old (not yet upgraded) nodes at the same time. This ensures that there are always at least 4 nodes available.

UsableSubnetwork
UsableSubnetwork resource returns the subnetwork name, its associated network and the primary CIDR range.

UsableSubnetworkSecondaryRange
Secondary IP range of a usable subnetwork.

VerticalPodAutoscaling
VerticalPodAutoscaling contains global, per-cluster information required by Vertical Pod Autoscaler to automatically adjust the resources of pods controlled by it.

WorkloadIdentityConfig
Configuration for the use of Kubernetes Service Accounts in GCP IAM policies.

WorkloadMetadataConfig
WorkloadMetadataConfig defines the metadata configuration to expose to workloads on the node pool.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.