Struct google_container1::api::Cluster

pub struct Cluster {
    pub addons_config: Option<AddonsConfig>,
    pub authenticator_groups_config: Option<AuthenticatorGroupsConfig>,
    pub autopilot: Option<Autopilot>,
    pub autoscaling: Option<ClusterAutoscaling>,
    pub binary_authorization: Option<BinaryAuthorization>,
    pub cluster_ipv4_cidr: Option<String>,
    pub conditions: Option<Vec<StatusCondition>>,
    pub create_time: Option<String>,
    pub current_master_version: Option<String>,
    pub current_node_count: Option<i32>,
    pub current_node_version: Option<String>,
    pub database_encryption: Option<DatabaseEncryption>,
    pub default_max_pods_constraint: Option<MaxPodsConstraint>,
    pub description: Option<String>,
    pub enable_kubernetes_alpha: Option<bool>,
    pub enable_tpu: Option<bool>,
    pub endpoint: Option<String>,
    pub expire_time: Option<String>,
    pub initial_cluster_version: Option<String>,
    pub initial_node_count: Option<i32>,
    pub instance_group_urls: Option<Vec<String>>,
    pub ip_allocation_policy: Option<IPAllocationPolicy>,
    pub label_fingerprint: Option<String>,
    pub legacy_abac: Option<LegacyAbac>,
    pub location: Option<String>,
    pub locations: Option<Vec<String>>,
    pub logging_service: Option<String>,
    pub maintenance_policy: Option<MaintenancePolicy>,
    pub master_auth: Option<MasterAuth>,
    pub master_authorized_networks_config: Option<MasterAuthorizedNetworksConfig>,
    pub monitoring_service: Option<String>,
    pub name: Option<String>,
    pub network: Option<String>,
    pub network_config: Option<NetworkConfig>,
    pub network_policy: Option<NetworkPolicy>,
    pub node_config: Option<NodeConfig>,
    pub node_ipv4_cidr_size: Option<i32>,
    pub node_pools: Option<Vec<NodePool>>,
    pub notification_config: Option<NotificationConfig>,
    pub private_cluster_config: Option<PrivateClusterConfig>,
    pub release_channel: Option<ReleaseChannel>,
    pub resource_labels: Option<HashMap<String, String>>,
    pub resource_usage_export_config: Option<ResourceUsageExportConfig>,
    pub self_link: Option<String>,
    pub services_ipv4_cidr: Option<String>,
    pub shielded_nodes: Option<ShieldedNodes>,
    pub status: Option<String>,
    pub status_message: Option<String>,
    pub subnetwork: Option<String>,
    pub tpu_ipv4_cidr_block: Option<String>,
    pub vertical_pod_autoscaling: Option<VerticalPodAutoscaling>,
    pub workload_identity_config: Option<WorkloadIdentityConfig>,
    pub zone: Option<String>,
}

A Google Kubernetes Engine cluster.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

locations clusters get projects (response)

zones clusters get projects (response)

Fields
addons_config: Option<AddonsConfig>

Configurations for the various addons available to run in the cluster.

authenticator_groups_config: Option<AuthenticatorGroupsConfig>

Configuration controlling RBAC group membership information.

autopilot: Option<Autopilot>

Autopilot configuration for the cluster.

autoscaling: Option<ClusterAutoscaling>

Cluster-level autoscaling configuration.

binary_authorization: Option<BinaryAuthorization>

Configuration for Binary Authorization.

cluster_ipv4_cidr: Option<String>

The IP address range of the container pods in this cluster, in CIDR notation (e.g. 10.96.0.0/14). Leave blank to have one automatically chosen or specify a /14 block in 10.0.0.0/8.

conditions: Option<Vec<StatusCondition>>

Which conditions caused the current cluster state.

create_time: Option<String>

[Output only] The time the cluster was created, in RFC3339 text format.

current_master_version: Option<String>

[Output only] The current software version of the master endpoint.

current_node_count: Option<i32>

[Output only] The number of nodes currently in the cluster. Deprecated. Call Kubernetes API directly to retrieve node information.

current_node_version: Option<String>

[Output only] Deprecated, use NodePools.version instead. The current version of the node software components. If they are currently at multiple versions because they're in the process of being upgraded, this reflects the minimum version of all nodes.

database_encryption: Option<DatabaseEncryption>

Configuration of etcd encryption.

default_max_pods_constraint: Option<MaxPodsConstraint>

The default constraint on the maximum number of pods that can be run simultaneously on a node in the node pool of this cluster. Only honored if cluster created with IP Alias support.

description: Option<String>

An optional description of this cluster.

enable_kubernetes_alpha: Option<bool>

Kubernetes alpha features are enabled on this cluster. This includes alpha API groups (e.g. v1alpha1) and features that may not be production ready in the kubernetes version of the master and nodes. The cluster has no SLA for uptime and master/node upgrades are disabled. Alpha enabled clusters are automatically deleted thirty days after creation.

enable_tpu: Option<bool>

Enable the ability to use Cloud TPUs in this cluster.

endpoint: Option<String>

[Output only] The IP address of this cluster's master endpoint. The endpoint can be accessed from the internet at https://username:password@endpoint/. See the masterAuth property of this resource for username and password information.

expire_time: Option<String>

[Output only] The time the cluster will be automatically deleted in RFC3339 text format.

initial_cluster_version: Option<String>

The initial Kubernetes version for this cluster. Valid versions are those found in validMasterVersions returned by getServerConfig. The version can be upgraded over time; such upgrades are reflected in currentMasterVersion and currentNodeVersion. Users may specify either explicit versions offered by Kubernetes Engine or version aliases, which have the following behavior: - "latest": picks the highest valid Kubernetes version - "1.X": picks the highest valid patch+gke.N patch in the 1.X version - "1.X.Y": picks the highest valid gke.N patch in the 1.X.Y version - "1.X.Y-gke.N": picks an explicit Kubernetes version - "","-": picks the default Kubernetes version

initial_node_count: Option<i32>

The number of nodes to create in this cluster. You must ensure that your Compute Engine resource quota is sufficient for this number of instances. You must also have available firewall and routes quota. For requests, this field should only be used in lieu of a "node_pool" object, since this configuration (along with the "node_config") will be used to create a "NodePool" object with an auto-generated name. Do not use this and a node_pool at the same time. This field is deprecated, use node_pool.initial_node_count instead.

instance_group_urls: Option<Vec<String>>

Deprecated. Use node_pools.instance_group_urls.

ip_allocation_policy: Option<IPAllocationPolicy>

Configuration for cluster IP allocation.

label_fingerprint: Option<String>

The fingerprint of the set of labels for this cluster.

legacy_abac: Option<LegacyAbac>

Configuration for the legacy ABAC authorization mode.

location: Option<String>

[Output only] The name of the Google Compute Engine zone or region in which the cluster resides.

locations: Option<Vec<String>>

The list of Google Compute Engine zones in which the cluster's nodes should be located. This field provides a default value if NodePool.Locations are not specified during node pool creation. Warning: changing cluster locations will update the NodePool.Locations of all node pools and will result in nodes being added and/or removed.

logging_service: Option<String>

The logging service the cluster should use to write logs. Currently available options: * logging.googleapis.com/kubernetes - The Cloud Logging service with a Kubernetes-native resource model * logging.googleapis.com - The legacy Cloud Logging service (no longer available as of GKE 1.15). * none - no logs will be exported from the cluster. If left as an empty string,logging.googleapis.com/kubernetes will be used for GKE 1.14+ or logging.googleapis.com for earlier versions.

maintenance_policy: Option<MaintenancePolicy>

Configure the maintenance policy for this cluster.

master_auth: Option<MasterAuth>

The authentication information for accessing the master endpoint. If unspecified, the defaults are used: For clusters before v1.12, if master_auth is unspecified, username will be set to "admin", a random password will be generated, and a client certificate will be issued.

master_authorized_networks_config: Option<MasterAuthorizedNetworksConfig>

The configuration options for master authorized networks feature.

monitoring_service: Option<String>

The monitoring service the cluster should use to write metrics. Currently available options: * "monitoring.googleapis.com/kubernetes" - The Cloud Monitoring service with a Kubernetes-native resource model * monitoring.googleapis.com - The legacy Cloud Monitoring service (no longer available as of GKE 1.15). * none - No metrics will be exported from the cluster. If left as an empty string,monitoring.googleapis.com/kubernetes will be used for GKE 1.14+ or monitoring.googleapis.com for earlier versions.

name: Option<String>

The name of this cluster. The name must be unique within this project and location (e.g. zone or region), and can be up to 40 characters with the following restrictions: * Lowercase letters, numbers, and hyphens only. * Must start with a letter. * Must end with a number or a letter.

network: Option<String>

The name of the Google Compute Engine network to which the cluster is connected. If left unspecified, the default network will be used.

network_config: Option<NetworkConfig>

Configuration for cluster networking.

network_policy: Option<NetworkPolicy>

Configuration options for the NetworkPolicy feature.

node_config: Option<NodeConfig>

Parameters used in creating the cluster's nodes. For requests, this field should only be used in lieu of a "node_pool" object, since this configuration (along with the "initial_node_count") will be used to create a "NodePool" object with an auto-generated name. Do not use this and a node_pool at the same time. For responses, this field will be populated with the node configuration of the first node pool. (For configuration of each node pool, see node_pool.config) If unspecified, the defaults are used. This field is deprecated, use node_pool.config instead.

node_ipv4_cidr_size: Option<i32>

[Output only] The size of the address space on each node for hosting containers. This is provisioned from within the container_ipv4_cidr range. This field will only be set when cluster is in route-based network mode.

node_pools: Option<Vec<NodePool>>

The node pools associated with this cluster. This field should not be set if "node_config" or "initial_node_count" are specified.

notification_config: Option<NotificationConfig>

Notification configuration of the cluster.

private_cluster_config: Option<PrivateClusterConfig>

Configuration for private cluster.

release_channel: Option<ReleaseChannel>

Release channel configuration.

resource_labels: Option<HashMap<String, String>>

The resource labels for the cluster to use to annotate any related Google Compute Engine resources.

resource_usage_export_config: Option<ResourceUsageExportConfig>

Configuration for exporting resource usages. Resource usage export is disabled when this config is unspecified.

self_link: Option<String>

[Output only] Server-defined URL for the resource.

services_ipv4_cidr: Option<String>

[Output only] The IP address range of the Kubernetes services in this cluster, in CIDR notation (e.g. 1.2.3.4/29). Service addresses are typically put in the last /16 from the container CIDR.

shielded_nodes: Option<ShieldedNodes>

Shielded Nodes configuration.

status: Option<String>

[Output only] The current status of this cluster.

status_message: Option<String>

[Output only] Deprecated. Use conditions instead. Additional information about the current status of this cluster, if available.

subnetwork: Option<String>

The name of the Google Compute Engine subnetwork to which the cluster is connected.

tpu_ipv4_cidr_block: Option<String>

[Output only] The IP address range of the Cloud TPUs in this cluster, in CIDR notation (e.g. 1.2.3.4/29).

vertical_pod_autoscaling: Option<VerticalPodAutoscaling>

Cluster-level Vertical Pod Autoscaling configuration.

workload_identity_config: Option<WorkloadIdentityConfig>

Configuration for the use of Kubernetes Service Accounts in GCP IAM policies.

zone: Option<String>

[Output only] The name of the Google Compute Engine zone in which the cluster resides. This field is deprecated, use location instead.