Struct google_container1::Cluster

pub struct Cluster {
    pub status: Option<String>,
    pub node_ipv4_cidr_size: Option<i32>,
    pub description: Option<String>,
    pub current_node_version: Option<String>,
    pub current_master_version: Option<String>,
    pub master_auth: Option<MasterAuth>,
    pub initial_node_count: Option<i32>,
    pub monitoring_service: Option<String>,
    pub create_time: Option<String>,
    pub network: Option<String>,
    pub endpoint: Option<String>,
    pub name: Option<String>,
    pub logging_service: Option<String>,
    pub instance_group_urls: Option<Vec<String>>,
    pub initial_cluster_version: Option<String>,
    pub node_config: Option<NodeConfig>,
    pub self_link: Option<String>,
    pub cluster_ipv4_cidr: Option<String>,
    pub zone: Option<String>,
    pub status_message: Option<String>,
    pub services_ipv4_cidr: Option<String>,
}

A Google Container Engine cluster.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

zones clusters get projects (response)

Fields

status

[Output only] The current status of this cluster.

node_ipv4_cidr_size

[Output only] The size of the address space on each node for hosting containers. This is provisioned from within the container_ipv4_cidr range.

description

An optional description of this cluster.

current_node_version

[Output only] The current version of the node software components. If they are currently at different versions because they're in the process of being upgraded, this reflects the minimum version of any of them.

current_master_version

[Output only] The current software version of the master endpoint.

master_auth

The authentication information for accessing the master.

initial_node_count

The number of nodes to create in this cluster. You must ensure that your Compute Engine resource quota is sufficient for this number of instances. You must also have available firewall and routes quota.

monitoring_service

The monitoring service that the cluster should write metrics to. Currently available options: * "monitoring.googleapis.com" - the Google Cloud Monitoring service * "none" - no metrics will be exported from the cluster * "" - default value; the default is "monitoring.googleapis.com"

create_time

[Output only] The time the cluster was created, in RFC3339 text format.

network

The name of the Google Compute Engine network to which the cluster is connected. If left unspecified, the "default" network will be used.

endpoint

[Output only] The IP address of this cluster's Kubernetes master endpoint. The endpoint can be accessed from the internet at https://username:password@endpoint/. See the masterAuth property of this resource for username and password information.

name

The name of this cluster. The name must be unique within this project and zone, and can be up to 40 characters with the following restrictions: * Lowercase letters, numbers, and hyphens only. * Must start with a letter. * Must end with a number or a letter.

logging_service

The logging service that the cluster should write logs to. Currently available options: * "logging.googleapis.com" - the Google Cloud Logging service * "none" - no logs will be exported from the cluster * "" - default value; the default is "logging.googleapis.com"

instance_group_urls

[Output only] The resource URLs of instance groups associated with this cluster.

initial_cluster_version

[Output only] The software version of Kubernetes master and kubelets used in the cluster when it was first created. The version can be upgraded over time.

node_config

Parameters used in creating the cluster's nodes. See the descriptions of the child properties of nodeConfig. If unspecified, the defaults for all child properties are used.

self_link

[Output only] Server-defined URL for the resource.

cluster_ipv4_cidr

The IP address range of the container pods in this cluster, in CIDR notation (e.g. 10.96.0.0/14). Leave blank to have one automatically chosen or specify a /14 block in 10.0.0.0/8.

zone

[Output only] The name of the Google Compute Engine zone in which the cluster resides.

status_message

[Output only] Additional information about the current status of this cluster, if available.

services_ipv4_cidr

[Output only] The IP address range of the Kubernetes services in this cluster, in CIDR notation (e.g. 1.2.3.4/29). Service addresses are typically put in the last /16 from the container CIDR.