Struct google_container1::MasterAuth

pub struct MasterAuth {
    pub username: Option<String>,
    pub password: Option<String>,
    pub client_key: Option<String>,
    pub client_certificate: Option<String>,
    pub cluster_ca_certificate: Option<String>,
}

The authentication information for accessing the master endpoint. Authentication can be done using HTTP basic auth or using client certificates.

This type is not used in any activity, and only used as part of another schema.

Fields
username: Option<String>

The username to use for HTTP basic authentication to the master endpoint.

password: Option<String>

The password to use for HTTP basic authentication to the master endpoint. Because the master endpoint is open to the Internet, you should create a strong password.

client_key: Option<String>

[Output only] Base64-encoded private key used by clients to authenticate to the cluster endpoint.

client_certificate: Option<String>

[Output only] Base64-encoded public certificate used by clients to authenticate to the cluster endpoint.

cluster_ca_certificate: Option<String>

[Output only] Base64-encoded public certificate that is the root of trust for the cluster.