Struct google_container1::NodeConfig

pub struct NodeConfig {
    pub local_ssd_count: Option<i32>,
    pub machine_type: Option<String>,
    pub tags: Option<Vec<String>>,
    pub preemptible: Option<bool>,
    pub labels: Option<HashMap<String, String>>,
    pub service_account: Option<String>,
    pub taints: Option<Vec<NodeTaint>>,
    pub oauth_scopes: Option<Vec<String>>,
    pub disk_size_gb: Option<i32>,
    pub accelerators: Option<Vec<AcceleratorConfig>>,
    pub min_cpu_platform: Option<String>,
    pub disk_type: Option<String>,
    pub image_type: Option<String>,
    pub metadata: Option<HashMap<String, String>>,
}

Parameters that describe the nodes in a cluster.

This type is not used in any activity, and only used as part of another schema.

Fields
local_ssd_count: Option<i32>

The number of local SSD disks to be attached to the node.

The limit for this value is dependant upon the maximum number of disks available on a machine per zone. See: https://cloud.google.com/compute/docs/disks/local-ssd#local_ssd_limits for more information.

machine_type: Option<String>

The name of a Google Compute Engine machine type (e.g. n1-standard-1).

If unspecified, the default machine type is n1-standard-1.

tags: Option<Vec<String>>

The list of instance tags applied to all nodes. Tags are used to identify valid sources or targets for network firewalls and are specified by the client during cluster or node pool creation. Each tag within the list must comply with RFC1035.

preemptible: Option<bool>

Whether the nodes are created as preemptible VM instances. See: https://cloud.google.com/compute/docs/instances/preemptible for more information about preemptible VM instances.

labels: Option<HashMap<String, String>>

The map of Kubernetes labels (key/value pairs) to be applied to each node. These will added in addition to any default label(s) that Kubernetes may apply to the node. In case of conflict in label keys, the applied set may differ depending on the Kubernetes version -- it's best to assume the behavior is undefined and conflicts should be avoided. For more information, including usage and the valid values, see: https://kubernetes.io/docs/concepts/overview/working-with-objects/labels/

service_account: Option<String>

The Google Cloud Platform Service Account to be used by the node VMs. If no Service Account is specified, the "default" service account is used.

taints: Option<Vec<NodeTaint>>

List of kubernetes taints to be applied to each node.

For more information, including usage and the valid values, see: https://kubernetes.io/docs/concepts/configuration/taint-and-toleration/

oauth_scopes: Option<Vec<String>>

The set of Google API scopes to be made available on all of the node VMs under the "default" service account.

The following scopes are recommended, but not required, and by default are not included:

https://www.googleapis.com/auth/compute is required for mounting persistent storage on your nodes.

https://www.googleapis.com/auth/devstorage.read_only is required for communicating with gcr.io (the Google Container Registry).

If unspecified, no scopes are added, unless Cloud Logging or Cloud Monitoring are enabled, in which case their required scopes will be added.

disk_size_gb: Option<i32>

Size of the disk attached to each node, specified in GB. The smallest allowed disk size is 10GB.

If unspecified, the default disk size is 100GB.

accelerators: Option<Vec<AcceleratorConfig>>

A list of hardware accelerators to be attached to each node. See https://cloud.google.com/compute/docs/gpus for more information about support for GPUs.

min_cpu_platform: Option<String>

Minimum CPU platform to be used by this instance. The instance may be scheduled on the specified or newer CPU platform. Applicable values are the friendly names of CPU platforms, such as minCpuPlatform: "Intel Haswell" or minCpuPlatform: "Intel Sandy Bridge". For more information, read how to specify min CPU platform

disk_type: Option<String>

Type of the disk attached to each node (e.g. 'pd-standard' or 'pd-ssd')

If unspecified, the default disk type is 'pd-standard'

image_type: Option<String>

The image type to use for this node. Note that for a given image type, the latest version of it will be used.

metadata: Option<HashMap<String, String>>

The metadata key/value pairs assigned to instances in the cluster.

Keys must conform to the regexp [a-zA-Z0-9-_]+ and be less than 128 bytes in length. These are reflected as part of a URL in the metadata server. Additionally, to avoid ambiguity, keys must not conflict with any other metadata keys for the project or be one of the reserved keys: "cluster-location" "cluster-name" "cluster-uid" "configure-sh" "containerd-configure-sh" "enable-os-login" "gci-update-strategy" "gci-ensure-gke-docker" "instance-template" "kube-env" "startup-script" "user-data" "disable-address-manager" "windows-startup-script-ps1" "common-psm1" "k8s-node-setup-psm1" "install-ssh-psm1" "user-profile-psm1" "serial-port-logging-enable"

Values are free-form strings, and only have meaning as interpreted by the image running in the instance. The only restriction placed on them is that each value's size must be less than or equal to 32 KB.

The total size of all keys and values must be less than 512 KB.