Struct google_container1::NodePool

pub struct NodePool {
    pub status: Option<String>,
    pub self_link: Option<String>,
    pub name: Option<String>,
    pub instance_group_urls: Option<Vec<String>>,
    pub version: Option<String>,
    pub initial_node_count: Option<i32>,
    pub config: Option<NodeConfig>,
    pub status_message: Option<String>,
}

NodePool contains the name and configuration for a cluster's node pool. Node pools are a set of nodes (i.e. VM's), with a common configuration and specification, under the control of the cluster master. They may have a set of Kubernetes labels applied to them, which may be used to reference them during pod scheduling. They may also be resized up or down, to accommodate the workload.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

zones clusters node pools get projects (response)

Fields
status: Option<String>

The status of the nodes in this pool instance.

self_link: Option<String>

Server-defined URL for the resource.

name: Option<String>

The name of the node pool.

instance_group_urls: Option<Vec<String>>

[Output only] The resource URLs of instance groups associated with this node pool.

version: Option<String>

The version of the Kubernetes of this node.

initial_node_count: Option<i32>

The initial node count for the pool. You must ensure that your Compute Engine resource quota is sufficient for this number of instances. You must also have available firewall and routes quota.

config: Option<NodeConfig>

The node configuration of the pool.

status_message: Option<String>

[Output only] Additional information about the current status of this node pool instance, if available.