Struct google_container1::Cluster

pub struct Cluster {
    pub node_ipv4_cidr_size: Option<i32>,
    pub status: Option<String>,
    pub current_master_version: Option<String>,
    pub description: Option<String>,
    pub current_node_version: Option<String>,
    pub addons_config: Option<AddonsConfig>,
    pub locations: Option<Vec<String>>,
    pub master_auth: Option<MasterAuth>,
    pub initial_node_count: Option<i32>,
    pub node_pools: Option<Vec<NodePool>>,
    pub monitoring_service: Option<String>,
    pub create_time: Option<String>,
    pub name: Option<String>,
    pub endpoint: Option<String>,
    pub current_node_count: Option<i32>,
    pub network: Option<String>,
    pub logging_service: Option<String>,
    pub instance_group_urls: Option<Vec<String>>,
    pub initial_cluster_version: Option<String>,
    pub node_config: Option<NodeConfig>,
    pub self_link: Option<String>,
    pub cluster_ipv4_cidr: Option<String>,
    pub subnetwork: Option<String>,
    pub zone: Option<String>,
    pub status_message: Option<String>,
    pub services_ipv4_cidr: Option<String>,
}

A Google Container Engine cluster.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

zones clusters get projects (response)

Fields
node_ipv4_cidr_size: Option<i32>

[Output only] The size of the address space on each node for hosting containers. This is provisioned from within the container_ipv4_cidr range.

status: Option<String>

[Output only] The current status of this cluster.

current_master_version: Option<String>

[Output only] The current software version of the master endpoint.

description: Option<String>

An optional description of this cluster.

current_node_version: Option<String>

[Output only] The current version of the node software components. If they are currently at multiple versions because they're in the process of being upgraded, this reflects the minimum version of all nodes.

addons_config: Option<AddonsConfig>

Configurations for the various addons available to run in the cluster.

locations: Option<Vec<String>>

The list of Google Compute Engine locations in which the cluster's nodes should be located.

master_auth: Option<MasterAuth>

The authentication information for accessing the master endpoint.

initial_node_count: Option<i32>

The number of nodes to create in this cluster. You must ensure that your Compute Engine resource quota is sufficient for this number of instances. You must also have available firewall and routes quota. For requests, this field should only be used in lieu of a "node_pool" object, since this configuration (along with the "node_config") will be used to create a "NodePool" object with an auto-generated name. Do not use this and a node_pool at the same time.

node_pools: Option<Vec<NodePool>>

The node pools associated with this cluster. When creating a new cluster, only a single node pool should be specified. This field should not be set if "node_config" or "initial_node_count" are specified.

monitoring_service: Option<String>

The monitoring service the cluster should use to write metrics. Currently available options: * monitoring.googleapis.com - the Google Cloud Monitoring service. * none - no metrics will be exported from the cluster. * if left as an empty string, monitoring.googleapis.com will be used.

create_time: Option<String>

[Output only] The time the cluster was created, in RFC3339 text format.

name: Option<String>

The name of this cluster. The name must be unique within this project and zone, and can be up to 40 characters with the following restrictions: * Lowercase letters, numbers, and hyphens only. * Must start with a letter. * Must end with a number or a letter.

endpoint: Option<String>

[Output only] The IP address of this cluster's master endpoint. The endpoint can be accessed from the internet at https://username:password@endpoint/. See the masterAuth property of this resource for username and password information.

current_node_count: Option<i32>

[Output only] The number of nodes currently in the cluster.

network: Option<String>

The name of the Google Compute Engine network to which the cluster is connected. If left unspecified, the default network will be used.

logging_service: Option<String>

The logging service the cluster should use to write logs. Currently available options: * logging.googleapis.com - the Google Cloud Logging service. * none - no logs will be exported from the cluster. * if left as an empty string,logging.googleapis.com will be used.

instance_group_urls: Option<Vec<String>>

[Output only] The resource URLs of instance groups associated with this cluster.

initial_cluster_version: Option<String>

[Output only] The software version of the master endpoint and kubelets used in the cluster when it was first created. The version can be upgraded over time.

node_config: Option<NodeConfig>

Parameters used in creating the cluster's nodes. See nodeConfig for the description of its properties. For requests, this field should only be used in lieu of a "node_pool" object, since this configuration (along with the "initial_node_count") will be used to create a "NodePool" object with an auto-generated name. Do not use this and a node_pool at the same time. For responses, this field will be populated with the node configuration of the first node pool. If unspecified, the defaults are used.

self_link: Option<String>

[Output only] Server-defined URL for the resource.

cluster_ipv4_cidr: Option<String>

The IP address range of the container pods in this cluster, in CIDR notation (e.g. 10.96.0.0/14). Leave blank to have one automatically chosen or specify a /14 block in 10.0.0.0/8.

subnetwork: Option<String>

The name of the Google Compute Engine subnetwork to which the cluster is connected.

zone: Option<String>

[Output only] The name of the Google Compute Engine zone in which the cluster resides.

status_message: Option<String>

[Output only] Additional information about the current status of this cluster, if available.

services_ipv4_cidr: Option<String>

[Output only] The IP address range of the Kubernetes services in this cluster, in CIDR notation (e.g. 1.2.3.4/29). Service addresses are typically put in the last /16 from the container CIDR.