Struct google_container1::ProjectMethods

pub struct ProjectMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on project resources. It is not used directly, but through the Container hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_container1 as container1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use container1::Container;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Container::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `zones_clusters_create(...)`, `zones_clusters_delete(...)`, `zones_clusters_get(...)`, `zones_clusters_list(...)`, `zones_clusters_node_pools_create(...)`, `zones_clusters_node_pools_delete(...)`, `zones_clusters_node_pools_get(...)`, `zones_clusters_node_pools_list(...)`, `zones_clusters_update(...)`, `zones_get_serverconfig(...)`, `zones_operations_get(...)` and `zones_operations_list(...)`
// to build up your call.
let rb = hub.projects();

Methods

impl<'a, C, A> ProjectMethods<'a, C, A>

fn zones_clusters_get(&self, project_id: &str, zone: &str, cluster_id: &str) -> ProjectZoneClusterGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the details of a specific cluster.

Arguments

projectId - The Google Developers Console project ID or project number.

zone - The name of the Google Compute Engine zone in which the cluster resides.

clusterId - The name of the cluster to retrieve.

fn zones_clusters_node_pools_create(&self, request: CreateNodePoolRequest, project_id: &str, zone: &str, cluster_id: &str) -> ProjectZoneClusterNodePoolCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a node pool for a cluster.

Arguments

request - No description provided.

projectId - The Google Developers Console project ID or project number.

zone - The name of the Google Compute Engine zone in which the cluster resides.

clusterId - The name of the cluster.

fn zones_clusters_node_pools_get(&self, project_id: &str, zone: &str, cluster_id: &str, node_pool_id: &str) -> ProjectZoneClusterNodePoolGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves the node pool requested.

Arguments

projectId - The Google Developers Console project ID or project number.

zone - The name of the Google Compute Engine zone in which the cluster resides.

clusterId - The name of the cluster.

nodePoolId - The name of the node pool.

fn zones_clusters_delete(&self, project_id: &str, zone: &str, cluster_id: &str) -> ProjectZoneClusterDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the cluster, including the Kubernetes endpoint and all worker nodes. Firewalls and routes that were configured during cluster creation are also deleted. Other Google Compute Engine resources that might be in use by the cluster (e.g. load balancer resources) will not be deleted if they weren't present at the initial create time.

Arguments

projectId - The Google Developers Console project ID or project number.

zone - The name of the Google Compute Engine zone in which the cluster resides.

clusterId - The name of the cluster to delete.

fn zones_get_serverconfig(&self, project_id: &str, zone: &str) -> ProjectZoneGetServerconfigCall<'a, C, A>

Create a builder to help you perform the following task:

Returns configuration info about the Container Engine service.

Arguments

projectId - The Google Developers Console project ID or project number.

zone - The name of the Google Compute Engine zone to return operations for.

fn zones_clusters_node_pools_list(&self, project_id: &str, zone: &str, cluster_id: &str) -> ProjectZoneClusterNodePoolListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the node pools for a cluster.

Arguments

projectId - The Google Developers Console project ID or project number.

zone - The name of the Google Compute Engine zone in which the cluster resides.

clusterId - The name of the cluster.

fn zones_operations_get(&self, project_id: &str, zone: &str, operation_id: &str) -> ProjectZoneOperationGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the specified operation.

Arguments

projectId - The Google Developers Console project ID or project number.

zone - The name of the Google Compute Engine zone in which the cluster resides.

operationId - The server-assigned name of the operation.

fn zones_operations_list(&self, project_id: &str, zone: &str) -> ProjectZoneOperationListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all operations in a project in a specific zone or all zones.

Arguments

projectId - The Google Developers Console project ID or project number.

zone - The name of the Google Compute Engine zone to return operations for, or - for all zones.

fn zones_clusters_create(&self, request: CreateClusterRequest, project_id: &str, zone: &str) -> ProjectZoneClusterCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a cluster, consisting of the specified number and type of Google Compute Engine instances. By default, the cluster is created in the project's default network. One firewall is added for the cluster. After cluster creation, the cluster creates routes for each node to allow the containers on that node to communicate with all other instances in the cluster. Finally, an entry is added to the project's global metadata indicating which CIDR range is being used by the cluster.

Arguments

request - No description provided.

projectId - The Google Developers Console project ID or project number.

zone - The name of the Google Compute Engine zone in which the cluster resides.

fn zones_clusters_update(&self, request: UpdateClusterRequest, project_id: &str, zone: &str, cluster_id: &str) -> ProjectZoneClusterUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the settings of a specific cluster.

Arguments

request - No description provided.

projectId - The Google Developers Console project ID or project number.

zone - The name of the Google Compute Engine zone in which the cluster resides.

clusterId - The name of the cluster to upgrade.

fn zones_clusters_list(&self, project_id: &str, zone: &str) -> ProjectZoneClusterListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists all clusters owned by a project in either the specified zone or all zones.

Arguments

projectId - The Google Developers Console project ID or project number.

zone - The name of the Google Compute Engine zone in which the cluster resides, or "-" for all zones.

fn zones_clusters_node_pools_delete(&self, project_id: &str, zone: &str, cluster_id: &str, node_pool_id: &str) -> ProjectZoneClusterNodePoolDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a node pool from a cluster.

Arguments

projectId - The Google Developers Console project ID or project number.

zone - The name of the Google Compute Engine zone in which the cluster resides.

clusterId - The name of the cluster.

nodePoolId - The name of the node pool to delete.